requirements.txt
Flask==1.0.2
flask_cors==3.0.6
mysql_connector_repackaged==0.3.1
newspaper3k==0.2.6